+++ src/neuralmpcx/__init__.py
@@ -10,7 +10,7 @@ r"""Neural MPCX (neuralmpcx) is a Model Predictive Control toolkit with neural M
 ==================== ========================================================
 """
 
-__version__ = "3.1.2"
+__version__ = "3.1.3"
 
 __all__ = ["Nlp", "Solution"]